chore: Move FlowStatusSection to its own folder
